+++ homepage/homepage/app/examples/page.tsx
@@ -1,28 +1,20 @@
+import { ExampleCard } from "@/components/examples/ExampleCard";
+import { ExampleDemo } from "@/components/examples/ExampleDemo";
 import { ClerkFullLogo } from "@/components/icons/ClerkFullLogo";
 import { NextjsLogo } from "@/components/icons/NextjsLogo";
 import { ReactLogo } from "@/components/icons/ReactLogo";
 import { ReactNativeLogo } from "@/components/icons/ReactNativeLogo";
+import { SvelteLogo } from "@/components/icons/SvelteLogo";
 import { VueLogo } from "@/components/icons/VueLogo";
+import { Example, features, tech } from "@/lib/example";
+import { clsx } from "clsx";
 import { H2 } from "gcmp-design-system/src/app/components/atoms/Headings";
+import { Icon } from "gcmp-design-system/src/app/components/atoms/Icon";
 import { GappedGrid } from "gcmp-design-system/src/app/components/molecules/GappedGrid";
 import { HeroHeader } from "gcmp-design-system/src/app/components/molecules/HeroHeader";
 
-import {
-  Schema_ts as ImageUploadSchema,
-  ImageUpload_tsx,
-} from "@/codeSamples/examples/image-upload/src";
-import {
-  Schema_ts as ReactionsSchema,
-  ReactionsScreen_tsx,
-} from "@/codeSamples/examples/reactions/src";
-import { ExampleCard } from "@/components/examples/ExampleCard";
-import { ExampleDemo } from "@/components/examples/ExampleDemo";
-import { SvelteLogo } from "@/components/icons/SvelteLogo";
-import { Example, features, tech } from "@/lib/example";
-import { Icon } from "gcmp-design-system/src/app/components/atoms/Icon";
-
 const MockButton = ({ children }: { children: React.ReactNode }) => (
-  <p className="bg-blue-100 text-blue-800 py-1 p-2 rounded-full font-medium text-center text-xs">
+  <p className="bg-blue-100 text-blue-800 py-1 px-3 rounded-full font-medium  text-xs inline-flex items-center justify-center">
     {children}
   </p>
 );
@@ -86,6 +78,69 @@ const OnboardingIllustration = () => (
           {text}
         </div>
       ))}
+    </div>
+  </div>
+);
+
+const OrganizationRow = ({
+  name,
+  members,
+  bgClassName,
+  children,
+}: {
+  name: string;
+  members: number;
+  bgClassName: string;
+  children?: React.ReactNode;
+}) => {
+  return (
+    <div className="flex items-center py-2 px-3">
+      <span
+        className={clsx(
+          bgClassName,
+          "size-8 rounded text-white font-medium text-lg inline-flex items-center justify-center mr-3",
+        )}
+      >
+        {name[0]}
+      </span>
+
+      <div className="flex-1">
+        <p className="text-sm font-medium text-stone-900 dark:text-white">
+          {name}
+        </p>
+        <p className="text-xs text-stone-500 dark:text-stone-600">
+          {members} {members > 1 ? "members" : "member"}
+        </p>
+      </div>
+
+      {children}
+    </div>
+  );
+};
+
+const OrganizationIllustration = () => (
+  <div className="flex bg-stone-100 h-full flex-col items-center justify-center dark:bg-transparent">
+    <div className="mx-auto">
+      <OrganizationRow
+        name="Garden Computing"
+        members={7}
+        bgClassName="bg-green-600"
+      >
+        <Icon name="chevronDown" className="ml-6" />
+      </OrganizationRow>
+
+      <div className="mt-3 rounded border divide-y shadow-sm bg-white dark:bg-transparent">
+        <OrganizationRow
+          name="Friends"
+          members={5}
+          bgClassName="bg-orange-600"
+        />
+        <OrganizationRow
+          name="Alice's projects"
+          members={1}
+          bgClassName="bg-rose-600"
+        />
+      </div>
     </div>
   </div>
 );
@@ -270,6 +325,24 @@ const reactExamples: Example[] = [
     illustration: <ChatIllustration />,
   },
   {
+    name: "Image upload",
+    slug: "image-upload",
+    description: "Learn how to upload and delete images",
+    tech: [tech.react],
+    features: [features.imageUpload],
+    demoUrl: "https://image-upload-demo.jazz.tools",
+    illustration: <ImageUploadIllustration />,
+  },
+  {
+    name: "Reactions",
+    slug: "reactions",
+    description: "Collect and render reactions from multiple users.",
+    tech: [tech.react],
+    features: [features.coFeed],
+    demoUrl: "https://reactions-demo.jazz.tools",
+    illustration: <ReactionsIllustration />,
+  },
+  {
     name: "Rate my pet",
     slug: "pets",
     description:
@@ -344,6 +417,15 @@ const reactExamples: Example[] = [
     features: [features.imageUpload, features.inviteLink],
     illustration: <OnboardingIllustration />,
   },
+  {
+    name: "Organization/Team",
+    slug: "organization",
+    description:
+      "Collaborate with members of your organization through a shared CoMap acting as a root or main database",
+    tech: [tech.react],
+    features: [features.inviteLink],
+    illustration: <OrganizationIllustration />,
+  },
 ];
 
 const nextExamples: Example[] = [
@@ -397,49 +479,6 @@ const vueExamples: Example[] = [
     illustration: (
       <div className="h-full w-full bg-cover bg-[url('/todo.jpg')] bg-left-bottom"></div>
     ),
-  },
-];
-
-const demos = [
-  {
-    name: "Image upload",
-    slug: "image-upload",
-    description: "Learn how to upload and delete images",
-    tech: [tech.react],
-    features: [features.imageUpload],
-    demoUrl: "https://image-upload-demo.jazz.tools",
-    illustration: <ImageUploadIllustration />,
-    showDemo: true,
-    codeSamples: [
-      {
-        name: "image-upload.tsx",
-        content: <ImageUpload_tsx />,
-      },
-      {
-        name: "schema.ts",
-        content: <ImageUploadSchema />,
-      },
-    ],
-  },
-  {
-    name: "Reactions",
-    slug: "reactions",
-    description: "Collect and render reactions from multiple users.",
-    tech: [tech.react],
-    features: [features.coFeed],
-    demoUrl: "https://reactions-demo.jazz.tools",
-    illustration: <ReactionsIllustration />,
-    showDemo: true,
-    codeSamples: [
-      {
-        name: "reactions.tsx",
-        content: <ReactionsScreen_tsx />,
-      },
-      {
-        name: "schema.ts",
-        content: <ReactionsSchema />,
-      },
-    ],
   },
 ];
 
@@ -502,14 +541,6 @@ export default function Page() {
         title="Example apps"
         slogan="Find an example app with code most similar to what you want to build"
       />
-
-      <div className="grid gap-8 mb-12 lg:gap-12">
-        <h2 className="sr-only">Example apps with demo and code</h2>
-        {demos.map(
-          (demo) =>
-            demo.showDemo && <ExampleDemo key={demo.slug} example={demo} />,
-        )}
-      </div>
 
       <div className="grid gap-12 lg:gap-20">
         {categories.map((category) => (