+++ packages/jazz-tools/src/tests/coMap.test.ts
@@ -38,8 +38,6 @@ describe("Simple CoMap operations", async () => {
     crypto: Crypto,
   });
 
-  console.log("TestMap schema", TestMap.prototype._schema);
-
   const birthday = new Date();
 
   const map = TestMap.create(
@@ -66,6 +64,25 @@ describe("Simple CoMap operations", async () => {
     ]);
   });
 
+  test("Construction with an Account", () => {
+    const map = TestMap.create(
+      { color: "red", _height: 10, birthday: birthday },
+      me,
+    );
+
+    expect(map.color).toEqual("red");
+  });
+
+  test("Construction with a Group", () => {
+    const group = Group.create(me);
+    const map = TestMap.create(
+      { color: "red", _height: 10, birthday: birthday },
+      group,
+    );
+
+    expect(map.color).toEqual("red");
+  });
+
   test("Construction with too many things provided", () => {
     const mapWithExtra = TestMap.create(
       {
@@ -80,6 +97,13 @@ describe("Simple CoMap operations", async () => {
     );
 
     expect(mapWithExtra.color).toEqual("red");
+  });
+
+  test("Empty schema", () => {
+    const emptyMap = CoMap.create({}, { owner: me });
+
+    // @ts-expect-error
+    expect(emptyMap.color).toEqual(undefined);
   });
 
   describe("Mutation", () => {
@@ -468,10 +492,8 @@ describe("CoMap resolution", async () => {
     const queue = new cojsonInternals.Channel<TestMap>();
 
     TestMap.subscribe(map.id, meOnSecondPeer, {}, (subscribedMap) => {
-      console.log(
-        "subscribedMap.nested?.twiceNested?.taste",
-        subscribedMap.nested?.twiceNested?.taste,
-      );
+      // Read to property to trigger loading
+      subscribedMap.nested?.twiceNested?.taste;
       void queue.push(subscribedMap);
     });